add solution for day 5

+++ common/utils.go
@@ -5,11 +5,11 @@ import (
 	"strconv"
 )
 
-func ParseInt(s string) int64 {
+func ParseInt(s string) int {
 	n, err := strconv.ParseInt(s, 10, 64)
 	if err != nil {
 		panic(fmt.Errorf("failed to parse %s as uint64: %w", s, err))
 	}
 
-	return n
+	return int(n)
 }
